+++ src/browser/extension/popup/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import Root from '../../../app/containers/Root';
+import Root from '../../../app/containers/App';
 import configureStore from '../../../app/store/configureStore';
 
 configureStore(store => {
