+++ src/browser/window/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
-import Root from '../../app/containers/Root';
+import Root from '../../app/containers/App';
 import configureStore from '../../app/store/configureStore';
 
 configureStore(store => {
